Move Util file to connectors since it contains connectors func
