+++ src/scraparr/connectors/util.py
